Move getContentUrlFromCID and revokeContentUrl to useContenUrl
